crecord: Move wave data structures to their own header
This is in preparation for adding wave parsing routines
to be used by tinycompress utilities.
